requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.42.0
sqlalchemy==2.0.48
psycopg2-binary==2.9.11
pydantic==2.12.5
python-multipart==0.0.22
celery==5.6.3
redis==7.4.0
pandas==3.0.2
docling==2.83.0
pytest==9.0.2
pyjwt==2.9.0
cryptography==42.0.8
